tap: fix tapping state machine
A bunch of states prematurely ended in IDLE which would result in single-tap
events on e.g. a two-finger hold past the timeout followed by a third finger
tap. Fix the state machine and update the diagram.

There's one final state DEAD which essentially represents a state were we
don't have anything to go on any further and need to be finger-counting until
we're back to IDLE.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>