Support finger pinning
On a clickpad, one finger may be used to press the physical button while the
other finger is moving around for drag-n-drop.

If a physical button event happens, pin the finger that's pressing the button
so we don't register movement from that finger until the button is released.
If the selected finger is the pointer finger, swap that to the next finger we
find.

Selection process is currently simple: the button pressing finger is the one
closer to the bottom edge of the touchpad.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>