Add a "now" argument to register timer
This is not quite ideal, but it allows the backend to calculate the time
offset (if any) between libtouchpad and the backend. In the case of the X
server, it may use CLOCK_MONOTONIC when libtouchpad merely uses whatever
comes through with the events.

A better solution is to set the device directly to CLOCK_MONOTONIC through
libevdev_set_clock_id().

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>