Drop the separate timer expiry function
Pass the current time into touchpad_handle_events() instead and let
libtouchpad figure out the rest. Previously we didn't handle any events off
the wire when handling time expiry but that leads to race conditions: if
events are sitting on the wire that happened before the timer expiry but we
the backend didn't call earlier, we might handle an expiry when we really
should be handling the events.

This patch doesn't fix this yet, it merely enables the fix for it.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>