Add support for software right-button click
This adds another state machine per-touch to detect what software-emulated
button we need to trigger when the touchpad physically clicks.

The touchpad is divided into three areas: left, right and "area". The buttons
are always configured as horizontal bar, divided into left/right. Anything
outside is "area" and defaults to a left button click.

If a finger is in the right button area when a click happens, a right click is
generated instead.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>