Change the test-suite to automatically test multiple devices
We only have one device defined so far but anyway. The last argument to
tptest_add takes a mask of devices, the test fixtures. The framework creates a
separate test case for each device, setting up the device as required before
the test.

This elevates the test case names to test suite names, so we'll have to be
more creative here

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>